LectureWithReminderOption domainmodel 패키지로 이동
