+++ app/src/main/java/com/wafflestudio/snutt2/data/tables/TableRepository.kt
@@ -1,9 +1,9 @@
 package com.wafflestudio.snutt2.data.tables
 
+import com.wafflestudio.snutt2.domainmodel.LectureWithReminderOption
 import com.wafflestudio.snutt2.lib.network.Result
 import com.wafflestudio.snutt2.lib.network.dto.core.SimpleTableDto
 import com.wafflestudio.snutt2.lib.network.dto.core.TableDto
-import com.wafflestudio.snutt2.views.logged_in.home.settings.LectureWithReminderOption
 import kotlinx.coroutines.flow.StateFlow
 
 interface TableRepository {
