+++ app/src/main/java/com/wafflestudio/snutt2/domainmodel/preview/PreviewData.kt
@@ -5,10 +5,10 @@ import com.wafflestudio.snutt2.lib.network.dto.core.NotificationDto
 import androidx.compose.ui.graphics.Color
 import com.wafflestudio.snutt2.domainmodel.CustomColor
 import com.wafflestudio.snutt2.domainmodel.LectureSession
+import com.wafflestudio.snutt2.domainmodel.LectureWithReminderOption
 import com.wafflestudio.snutt2.domainmodel.SyllabusLecture
 import com.wafflestudio.snutt2.domainmodel.domainModel
 import com.wafflestudio.snutt2.views.logged_in.home.settings.LectureReminderOffset
-import com.wafflestudio.snutt2.views.logged_in.home.settings.LectureWithReminderOption
 import java.time.DayOfWeek
 import java.time.LocalTime
