+++ app/src/main/java/com/wafflestudio/snutt2/lib/network/dto/core/TimetableLectureReminderDto.kt
@@ -2,8 +2,8 @@ package com.wafflestudio.snutt2.lib.network.dto.core
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
+import com.wafflestudio.snutt2.domainmodel.LectureWithReminderOption
 import com.wafflestudio.snutt2.views.logged_in.home.settings.LectureReminderOffset
-import com.wafflestudio.snutt2.views.logged_in.home.settings.LectureWithReminderOption
 
 // NOTE: 강의 리마인더가 지정된 적 없는 강의의 경우 empty response가 오는데 그 경우 EOFException이 나게 되고, ExceptionMapper에서 DomainError.EOF로 변환한다.
 @JsonClass(generateAdapter = true)
